Move forwarded_storage proto into its own file under the loadtest package rename cue file for panelmatch emulators (remove storage)
