+++ src/main/kotlin/org/wfanet/panelmatch/client/storage/forwarded/BUILD.bazel
@@ -1,12 +1,15 @@
 load("@io_bazel_rules_kotlin//kotlin:jvm.bzl", "kt_jvm_library")
 
-package(default_visibility = ["//visibility:public"])
+package(default_testonly = True,
+        default_visibility = ["//visibility:public"]
+)
 
 kt_jvm_library(
     name = "forwarded_storage",
     srcs = glob(["*.kt"]),
     deps = [
         "//src/main/kotlin/org/wfanet/panelmatch/client/storage",
+        "//src/main/proto/wfa/panelmatch/client/loadtest:forwarded_storage_kt_jvm_proto",
         "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/common/grpc",
         "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/storage/forwarded",
     ],
