+++ src/main/kotlin/org/wfanet/panelmatch/client/storage/forwarded/ForwardedStorageFactory.kt
@@ -21,7 +21,7 @@ import org.wfanet.measurement.internal.testing.ForwardedStorageGrpcKt
 import org.wfanet.measurement.storage.StorageClient
 import org.wfanet.measurement.storage.forwarded.ForwardedStorageClient
 import org.wfanet.panelmatch.client.storage.StorageDetails
-import org.wfanet.panelmatch.client.storage.StorageDetails.ForwardedStorage
+import org.wfanet.panelmatch.client.loadtest.ForwardedStorage
 import org.wfanet.panelmatch.common.ExchangeDateKey
 import org.wfanet.panelmatch.common.storage.StorageFactory
