+++ src/test/kotlin/org/wfanet/panelmatch/integration/k8s/BUILD.bazel
@@ -40,6 +40,7 @@ kt_jvm_library(
         "//src/main/proto/wfa/measurement/api/v2alpha:exchanges_service_kt_jvm_proto",
         "//src/main/proto/wfa/panelmatch/client/storage:storage_details_java_proto",
         "//src/main/proto/wfa/panelmatch/client/storage:storage_details_kt_jvm_proto",
+        "//src/main/proto/wfa/panelmatch/client/loadtest:forwarded_storage_kt_jvm_proto",
         "@wfa_common_jvm//imports/java/com/google/common/truth",
         "@wfa_common_jvm//imports/java/com/google/common/truth/extensions/proto",
         "@wfa_common_jvm//imports/java/com/google/protobuf",
